Change route of playlist application
player routes are now on /playlist/device
playlist entries are now on /playlist/entries
