+++ dakara_server/playlist/views_device.py
